Moved serializer to top level
