+++ MidsReborn/Forms/OptionsMenuItems/DbEditor/frmDBConvert.cs
@@ -1,21 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
 using System.IO;
 using System.Linq;
-using System.Reflection;
-using System.Security.Cryptography;
-using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows.Forms;
-using Mids_Reborn.My;
 using mrbBase;
 using mrbBase.Base.Master_Classes;
-using Syncfusion.Styles;
-using Syncfusion.Windows.Forms.Tools;
 using WK.Libraries.BetterFolderBrowserNS;
 
 namespace Mids_Reborn.Forms.OptionsMenuItems.DbEditor
@@ -218,7 +208,7 @@ namespace Mids_Reborn.Forms.OptionsMenuItems.DbEditor
             await Task.Delay(500);
             foreach (var power in DatabaseAPI.Database.Power) power.BaseRechargeTime = power.RechargeTime;
             Array.Sort(DatabaseAPI.Database.Power);
-            var serializer = MyApplication.GetSerializer();
+            var serializer = Serializer.GetSerializer();
             statusText.Items.Add("Assigning static indexes...");
             DatabaseAPI.AssignStaticIndexValues(serializer, false);
             DatabaseAPI.MatchAllIDs();
