+++ MidsReborn/Forms/OptionsMenuItems/DbEditor/frmDBDiffing.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 using mrbBase;
