+++ MidsReborn/Forms/OptionsMenuItems/DbEditor/frmEditPower.cs
@@ -1,6 +1,5 @@
 using System;
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Drawing;
 using System.Globalization;
 using System.IO;
