+++ MidsReborn/Forms/OptionsMenuItems/DbEditor/frmSetListing.cs
@@ -5,7 +5,6 @@ using System.Drawing;
 using System.Globalization;
 using System.Runtime.CompilerServices;
 using System.Windows.Forms;
-using Mids_Reborn.My;
 using mrbBase;
 using mrbBase.Base.Display;
 using mrbBase.Base.Extensions;
@@ -181,7 +180,7 @@ namespace Mids_Reborn.Forms.OptionsMenuItems.DbEditor
 
         private void btnSave_Click(object sender, EventArgs e)
         {
-            var serializer = MyApplication.GetSerializer();
+            var serializer = Serializer.GetSerializer();
             DatabaseAPI.SaveEnhancementDb(serializer);
             Hide();
         }
