+++ MidsReborn/Forms/UpdateSystem/PatchNotes.cs
@@ -1,13 +1,10 @@
 ﻿using System;
-using System.Diagnostics;
 using System.Drawing;
 using System.IO;
-using System.Linq;
 using System.Net;
 using System.Text;
 using System.Text.RegularExpressions;
 using System.Windows.Forms;
-using mrbBase;
 using mrbBase.Base.Master_Classes;
 
 namespace Mids_Reborn.Forms.UpdateSystem
