+++ MidsReborn/Forms/WindowMenuItems/frmDPSCalc.cs
@@ -1,5 +1,4 @@
 using System;
-using System.ComponentModel;
 using System.Drawing;
 using System.Globalization;
 using System.Threading;
