+++ MidsReborn/Forms/frmBusy.cs
@@ -1,5 +1,4 @@
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Windows.Forms;
 
 namespace Mids_Reborn.Forms
