+++ MidsReborn/Forms/frmPrint.cs
@@ -1,6 +1,5 @@
 using System;
 using System.ComponentModel;
-using System.Diagnostics;
 using System.Drawing.Printing;
 using System.Windows.Forms;
 using mrbBase;
