+++ MidsReborn/MainModule.cs
@@ -1,11 +1,8 @@
 using System;
 using System.Drawing;
-using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using Mids_Reborn.Forms;
-using Mids_Reborn.Forms.OptionsMenuItems;
-using Mids_Reborn.Forms.OptionsMenuItems.DbEditor;
 using mrbBase;
 using mrbBase.Base.Master_Classes;
