+++ MidsReborn/Program.cs
@@ -3,7 +3,6 @@ using System.Diagnostics;
 using System.Linq;
 using System.Windows.Forms;
 using Mids_Reborn.Forms;
-using Mids_Reborn.UIv2;
 using mrbBase.Base.Master_Classes;
 
 namespace Mids_Reborn
