+++ MidsReborn/UIv2/v2Controls/PowersControl.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
 using System.Drawing;
-using System.Linq;
 using System.Windows.Forms;
-using System.Windows.Forms.VisualStyles;
 using mrbBase;
 
 namespace Mids_Reborn.UIv2.v2Controls
